pyproject.toml
[build-system]
requires = ["setuptools>=69.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xiaohongshu-blogger-crawler"
version = "0.1.0"
description = "Python crawler scaffold for collecting public Xiaohongshu blogger data."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
  { name = "Your Team" }
]
dependencies = [
  "fastapi>=0.111.0,<1.0.0",
  "httpx>=0.27.0,<1.0.0",
  "pydantic>=2.8.0,<3.0.0",
  "python-dotenv>=1.0.1,<2.0.0",
  "rich>=13.7.1,<14.0.0",
  "tenacity>=8.3.0,<9.0.0",
  "typer>=0.12.3,<1.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.10.0,<2.0.0",
  "pytest>=8.2.0,<9.0.0",
  "pytest-asyncio>=0.23.7,<1.0.0",
  "ruff>=0.5.0,<1.0.0",
]

[project.scripts]
xhs-crawler = "xiaohongshu_blogger_crawler.cli:app"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
packages = ["xiaohongshu_blogger_crawler"]
